requirements.txt
apache-beam[gcp]==2.48.0
google-cloud-bigquery
requests